Drag Pan Added
Cleaned up some code, added a nice drag pan, and I think the camera system is fairly complete at this point.  Only thing left to do for foreseeable future is to refactor this to the Unity Input system so it could be used across platforms.....but I think we might have to rethink the control scheme for a controller anyway.